guest_tools/win: Split destructive tests
Selecting TestGuestToolsWindowsDestructive for a test subset is tricky,
and splitting it into a new dedicated file will make that easier.
